Sixteen more plans, and the first rider whose host premium is published
Pacific Cross ×5 swept from its own anonymous quote endpoint, Navakij ×5 from
a complete 13-band table, Generali ×3 from a 14-band sex-split brochure table,
and Ocean Life ×3.

Ocean Life OCHI is the one that matters. Its checkout itemises the host
whole-life policy separately at every age for both sexes, so the share of the
premium that actually buys health cover is a real number here for the first
time — 73.4% at 35. Every other rider in the dataset still returns null there,
because no other insurer publishes what the compulsory host policy costs.

Nine records now carry a premium for every year from entry to the renewal
ceiling, up from two. Seven cite filed wording, up from two. Sixteen carry an
exclusion list, and Ocean Life's comes from the contract rather than a
brochure.

Where an insurer publishes only part of the truth, the record says so instead
of filling the gap: Generali never names or prices the main policy its own
brochure says the rider dies with; Pacific Cross publishes its wording in
English only, under a footer admitting the unpublished Thai version prevails;
Navakij publishes no total inpatient ceiling at all, only per-category
sub-limits.

thb_per_year drops its integer constraint. Ocean Life quotes host premiums in
half-baht, and rounding 130 published figures to fit the schema would be this
project editing an insurer's numbers.

+++ src/lib/schema.ts
@@ -26,7 +26,12 @@ const Sourced = z.object({
 const PremiumBand = z.object({
 	age_from: z.number().int().min(0).max(120),
 	age_to: z.number().int().min(0).max(120),
-	thb_per_year: z.number().int().positive(),
+	/**
+	 * Not an integer: Ocean Life quotes host premiums in half-baht (720.50), and
+	 * rounding a published figure to make it fit the schema would be this project
+	 * editing an insurer's number. Display already formats to whole baht.
+	 */
+	thb_per_year: z.number().positive(),
 	sex: z.enum(['any', 'male', 'female']).default('any')
 });
